change the name of screenshot and finish unittest

+++ Week2/PracB/BankTask2/BankTask2.cs
@@ -44,7 +44,7 @@ namespace Bank
                 throw new ArgumentException("Amount can not be negative.");
                 
             }
-            result=(decimal)Balance+amount;
+            result=Balance+amount;
             Balance=result;
             return Balance;
